requirements.txt
Pillow==9.4.0
pillow_heif==0.9.2
tqdm==4.64.1
